pyproject.toml
[project]
name = "chatbot-project-llm"
version = "0.1.0"
description = "RAG pipeline for Vietnamese legal QA assignment"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.5",
  "tqdm>=4.66",
  "faiss-cpu>=1.8.0.post1",
  "opensearch-py>=2.6",
  "openpyxl>=3.1",
  "pypdf>=5.0",
  "python-dotenv>=1.0",
  "pyarrow>=23.0.1",
]

[dependency-groups]
embed = [
  "sentence-transformers>=3.0",
  "transformers>=4.44",
  "torch>=2.4",
]
llm = [
  "langchain>=0.3",
  "langchain-community>=0.3",
  "langchain-text-splitters>=0.3",
  "langchain-huggingface>=0.1",
]
dev = [
  "pytest>=8.3",
  "ruff>=0.6",
  "ipykernel>=6.29",
]

[tool.uv]
package = false
